Updated logic for deleteOutputContents to test the path to ensure that it's valid before it tries to delete the files in the directory.

+++ Scripts/Export modules from Excel.ps1
@@ -52,7 +52,7 @@ function Export-ExcelModules {
                 $TypesArr += [VbaType]::userform
             }
     
-            if ($deleteOutputContents) {
+            if ($deleteOutputContents -and (Test-Path -Path $outputPath)) {
                 Remove-Item -Path "$outputPath\*" -Recurse
             }
